requirements.txt
colorama~=0.4.4

mysql-connector-python~=8.1.0
requests~=2.25.1

bs4~=0.0.1
beautifulsoup4~=4.12.2
lxml
pdfquery~=0.4.3